fix(Megamenu): add margin to prevent shifting on scroll lock (#5926)
### Related Ticket(s)

Masthead Navigation moving on toggle #5472

### Description

For users on windows machines or have set the scrollbar to always show on their Macs, the content and masthead will shift when opening and closing the megamenu. This is due to the scrollbar lock when the megamenu is open, removing the scrollbar. The content then shifts due to the scrollbar disappearing.

Solution is to get the width of the scrollbar and then add the width to the body and masthead as `margin-right`.

BEFORE:
![before](https://user-images.githubusercontent.com/54281166/116476514-5362e080-a849-11eb-95a9-81c9862d24c5.gif)

AFTER:
![after](https://user-images.githubusercontent.com/54281166/116476523-565dd100-a849-11eb-9caa-81743ab25bb3.gif)

### Changelog

**Changed**

- make sure the `body` and masthead add a margin-right of the scrollbar width when the megamenu is expanded to prevent content from shifting

<!-- React and Web Component deploy previews are enabled by default. -->
<!-- To enable additional available deploy previews, apply the following -->
<!-- labels for the corresponding package: -->
<!-- *** "package: services": Services -->
<!-- *** "package: utilities": Utilities -->
<!-- *** "package: styles": Carbon Expressive -->
<!-- *** "RTL": React / Web Components (RTL) -->
<!-- *** "feature flag": React / Web Components (experimental) -->