fix(masthead-search): disable focus for search open on load (#5971)
### Related Ticket(s)
#5924

### Description
Due to the way the `MastheadSearch` and `MastheadSearchInput` components are built in React, whenever the Search bar is open it's always assumed it's open state came from clicking the Search button to bring up the searchbar. This causes for the focus state go into the bar the moment the "Search Open on Load" story loads. 

This PR adds an additional check before the focus function is called in case `disableFocus` is set to `true`. After `MastheadSearchInput` closes, this boolean is changed to `false` in order for the focus feature to activate as normal upon reopen.

Note that using boolean React states for this would need a bigger refactoring, since whenever `MastheadSearchInput` is closed, its boolean values coming from `MastheadSearch` are reset to what they originally were, meaning that if `disableFocus` was set to `false` upon closing, it would reset back to `true` the next time it opens. With the current fix this boolean is only set to `true` in the "Search Open on Load" story.

### Changelog

**New**

- `disableFocus` prop in `MastheadSearchInput`

<!-- React and Web Component deploy previews are enabled by default. -->
<!-- To enable additional available deploy previews, apply the following -->
<!-- labels for the corresponding package: -->
<!-- *** "package: services": Services -->
<!-- *** "package: utilities": Utilities -->
<!-- *** "package: styles": Carbon Expressive -->
<!-- *** "RTL": React / Web Components (RTL) -->
<!-- *** "feature flag": React / Web Components (experimental) -->