fix(card-group): card grid alignment (#5974)
### Related Ticket(s)

https://github.com/carbon-design-system/carbon-for-ibm-dotcom/issues/5730

### Description

@shixiedesign I can fix the alignment perfectly when the card only includes text:

![image](https://user-images.githubusercontent.com/4054756/117249083-3b1d3380-adfe-11eb-9778-6cb63f3b174a.png)

The problem happens when the card includes an image. From my tests, it's mathematically impossible to make it work unless we add a javascript fix that makes all images inside the card same height. To make the last card's content align with the grid, it needs to be the same width as the top card, which takes 33.333...% of the width. If you substract that width plus the two 16px gutters from 100%, we don't have enough space left to make the other two cards be the same row the same width as the last card, which causes their images to be shorter than the last card. 

![image](https://user-images.githubusercontent.com/4054756/117249312-9e0eca80-adfe-11eb-8b41-db82671907d9.png)

I can fix the image height problem by making them all the same width, but the last card wouldn't hang like the top card and image:

![image](https://user-images.githubusercontent.com/4054756/117249192-6acc3b80-adfe-11eb-8256-44deffab0f9a.png)

Not sure what to do? 

<!-- React and Web Component deploy previews are enabled by default. -->
<!-- To enable additional available deploy previews, apply the following -->
<!-- labels for the corresponding package: -->
<!-- *** "package: services": Services -->
<!-- *** "package: utilities": Utilities -->
<!-- *** "package: styles": Carbon Expressive -->
<!-- *** "RTL": React / Web Components (RTL) -->
<!-- *** "feature flag": React / Web Components (experimental) -->