fix(lightbox-video-player): add check for video cta specific modal to open and prevent multiple modals from opening (#6260)
### Related Ticket(s)

Multiple lightboxes appear when using both the VideoPlayerContainer and VideoCTAContainer #6254

### Description

Issue 1:
`video-cta-container` renders the `lightbox-video-player-composite` component and triggers the opening of the modal itself. When placed with `video-player-container` that requires `lightbox-video-player-container` to be applied as well, the custom event triggers both lightbox modals (one for the `video-player-container` and one for `video-cta-container`) to open.

Solution is to pass a `video-cta-lightbox` attribute to `lightbox-video-player-composite` within `video-cta-composite` and check whether to open the lightbox modal based off the attribute. 

Issue 2:
When having multiple video players that open to the lightbox, after opening each one and closing if user opens the first video again,  the video is that of the previously closed.

Solution is to set the hidden state to previously closed videos. Looks like there was code to do this in the `video-player-composite` that wasn't working properly. That logic was tweaked and moved to the `lightbox-video-player-composite` instead.

```
<dds-video-cta-container>
   <dds-text-cta cta-type="video" href="1_8ij4i1qy">
        test 2
    </dds-text-cta>
</dds-video-cta-container>
<dds-video-player-container video-id="1_9h94wo6b" aspect-ratio="4x3" playing-mode="lightbox"> </dds-video-player-container>
<dds-video-player-container video-id="1_w5drm2rz" aspect-ratio="4x3" playing-mode="lightbox"> </dds-video-player-container>
<dds-lightbox-video-player-container></dds-lightbox-video-player-container>
```

### Changelog

**New**

- pass `video-cta-lightbox` attribute to `lightbox-video-player-composite` to identify video-cta specific lightbox modal

**Changed**

- place hiding video logic into `lightbox-video-box-composite` instead

**Removed**

- hiding video logic from `video-player-composite`

<!-- React and Web Component deploy previews are enabled by default. -->
<!-- To enable additional available deploy previews, apply the following -->
<!-- labels for the corresponding package: -->
<!-- *** "package: services": Services -->
<!-- *** "package: utilities": Utilities -->
<!-- *** "package: styles": Carbon Expressive -->
<!-- *** "RTL": React / Web Components (RTL) -->
<!-- *** "feature flag": React / Web Components (experimental) -->