fix(e2e): adding timeout for render time on CI
Our CI is taking some time to render the pages before testing, which is
causing some timeout errors. This will bump it up to give our CI more
time for testing.

In addition, video player visual snapshots are showing false positives,
so suppressing the percy snapshots for those and only checking for build
 errors for now.