fix(styles): temporarily hard-code breakpoint values (#6557)
### Related Ticket(s)

#6554 

### Description

Temporarily use numerical breakpoint values for the `carbon--breakpoint-between` mixin. Currently there is an error in this mixin that prevents the usage of breakpoint keywords, i.e. `md`, `lg`, etc.

A [PR](https://github.com/carbon-design-system/carbon/pull/9102) for this fix is currently open. Once merged these changes can be reverted to use keywords.

### To-do
Revert these changes when the changes are available in the next `carbon-components` release. [#6558]

### Changelog

**Changed**

- use numerical values instead of keywords for `carbon--breakpoint-between` mixin

<!-- React and Web Component deploy previews are enabled by default. -->
<!-- To enable additional available deploy previews, apply the following -->
<!-- labels for the corresponding package: -->
<!-- *** "package: services": Services -->
<!-- *** "package: utilities": Utilities -->
<!-- *** "package: styles": Carbon Expressive -->
<!-- *** "RTL": React / Web Components (RTL) -->
<!-- *** "feature flag": React / Web Components (experimental) -->