test(e2e): creating dynamically generated index file (#6566)
### Related Ticket(s)

No related issue

### Description

This adds a dynamically created index file for the deploy preview of the web components codesandbox build, so that reviewers can view the built files. Currently there isn't an index file, so it would be tough for a review to navigate the deploy preview instance. It's no frills, but the output looks like this:

![image](https://user-images.githubusercontent.com/1641214/124607201-9dfd6c00-de3b-11eb-885b-357d0cd09f45.png)

### Changelog

**Changed**

- Adding new method to auto-create the index file based on the examples folders