chore(styles): update clean-css to remove duplicates and includes (#6609)
### Description

This fixes an issue where an unneeded style reset was included in the global imports file. This resulted in hundreds of duplicated blocks in the outputted files found in `/dist`, as well as styles compiled by bundlers such as Webpack. As this reset is included in `carbon-components/scss/globals/scss/css--reset` (which is already imported globally), there is no need to call it again.

Lots and lots (and lots!) of these – among others, but mostly these.

![image](https://user-images.githubusercontent.com/909118/125318021-7872d580-e307-11eb-8a53-f855359d14f2.gif)

Removing this, along with modifying the optimizer used in Gulp to remove duplicates and comments, has reduced the file size of unminified css files by ~3mb and removed ~80k lines of redundant/unneeded code.

**Before**
![Screen-Shot-2021-07-12-at-10 59 28-AM](https://user-images.githubusercontent.com/909118/125319256-95f46f00-e308-11eb-9554-9eb7928facb6.gif)
![Screen Shot 2021-07-12 at 11 07 00 AM](https://user-images.githubusercontent.com/909118/125318836-38f8b900-e308-11eb-89bd-acd6aa0489d9.png)


**After**
![Screen-Shot-2021-07-12-at-11 00 42-AM](https://user-images.githubusercontent.com/909118/125319284-9e4caa00-e308-11eb-9708-5131cc5302b2.gif)
![Screen Shot 2021-07-12 at 11 02 26 AM](https://user-images.githubusercontent.com/909118/125318926-4e6de300-e308-11eb-87da-7cb03013b178.png)



### Changelog

**New**

- add optimization levels 1 (remove comments) and 2 (remove duplicate rules) to `gulp-clean-css`

**Removed**

- `@include carbon--type-reset()` removed from global imports

<!-- React and Web Component deploy previews are enabled by default. -->
<!-- To enable additional available deploy previews, apply the following -->
<!-- labels for the corresponding package: -->
<!-- *** "package: services": Services -->
<!-- *** "package: utilities": Utilities -->
<!-- *** "package: styles": Carbon Expressive -->
<!-- *** "RTL": React / Web Components (RTL) -->
<!-- *** "feature flag": React / Web Components (experimental) -->