fix(locale-modal): set locale modal to not disappear on mobile keyboard (#6628)
### Related Ticket(s)

Web component: Dotcom shell - Language selection modal - user clicks on the modal, focus comes back to the parent page #6076

### Description

This issue appears on the DotcomShell story when keyboard shows up on mobile, it causes the page to scroll to focus on the input field. This triggers the Table of Contents scrolling logic and causes the modal to close. Preventing the triggering of the Table of Contents logic resolves this.

BEFORE:
https://user-images.githubusercontent.com/54281166/125505498-7b2c1bcd-7046-4e5b-b71b-1b6e113e4bbf.mp4

AFTER:

https://user-images.githubusercontent.com/54281166/125506469-ca03c279-0e19-4db1-9d78-523e8a34d629.mp4

### Changelog

**Changed**

- trigger table of contents logic when locale modal isn't open.
- lock page scrolling on opening of locale modal.

<!-- React and Web Component deploy previews are enabled by default. -->
<!-- To enable additional available deploy previews, apply the following -->
<!-- labels for the corresponding package: -->
<!-- *** "package: services": Services -->
<!-- *** "package: utilities": Utilities -->
<!-- *** "package: styles": Carbon Expressive -->
<!-- *** "RTL": React / Web Components (RTL) -->
<!-- *** "feature flag": React / Web Components (experimental) -->