fix(locale-modal): set region item to be tabbable for voiceover (#6735)
### Related Ticket(s)

Web component: Dotcom shell - Locale modal- when we activate "Select geographic area" link, the First modal is not displayed #6077

### Description

When using voiceover on iphone, the user isn't taken back to the first modal with all the geographic regions (Americas, Asia Pacific, etc) after selecting the "Select geographic area" link to go back. The "Americas" region card is briefly focused before voiceover announces the `select` component. 

Solution is the set the `tabIndex` of the first `<dds-region-item>` component in the first modal. This way voiceover recognizes the component as tabbable and will then focus properly on it.

### Changelog

**Changed**

- set `tabIndex` of the first region item to 0 to ensure voiceover focuses on it

<!-- React and Web Component deploy previews are enabled by default. -->
<!-- To enable additional available deploy previews, apply the following -->
<!-- labels for the corresponding package: -->
<!-- *** "package: services": Services -->
<!-- *** "package: utilities": Utilities -->
<!-- *** "package: styles": Carbon Expressive -->
<!-- *** "RTL": React / Web Components (RTL) -->
<!-- *** "feature flag": React / Web Components (experimental) -->