feat(loadPlex): introduce loadPlex utility
This utility will dynamically load in non-Latin fonts, as well as
setting the font-family to the page accordingly. This leverages the
existing Plex CDN fonts.

In addition, the globalInit introduced the execution of this code, so
this ensures that any page that loads in the Masthead or Footer will do
this check and load in the fonts accordingly.