feat(search-with-typeahead): introducing component (#6731)
### Related Ticket(s)
#6143 
#6144

### Description
The current Masthead uses Redux actions in order to incorporate the SearchAPI functionality into `dds-masthead-search`. This includes having a container, composite, and a connect component to render the Masthead Search component, as well as including these actions in the Masthead container and composite components. 

This PR extracts all the functionality from `<dds-masthead-search>` and ensures all the Search API and suggestions functionality can work on its own, meaning that we now effectively have a Search component with Typeahead functionality. 

This PR also replaces the aforementioned Masthead Search components within the actual Masthead. 

### Changelog

**New**

- extracted DDSMastheadSearch functionality into its own component DDSSearchWithTypeahead
- DDSSearchWIthTypeahead React wrapper

**Changed**

- replaced all the DDSMastheadSearch components with the new DDSSearchWithTypeahead component from the Masthead
- refactored some of the search styles

**Removed**

- DDSMastheadSearch, Connect, Container, Composite files

<!-- React and Web Component deploy previews are enabled by default. -->
<!-- To enable additional available deploy previews, apply the following -->
<!-- labels for the corresponding package: -->
<!-- *** "package: services": Services -->
<!-- *** "package: utilities": Utilities -->
<!-- *** "package: styles": Carbon Expressive -->
<!-- *** "RTL": React / Web Components (RTL) -->
<!-- *** "feature flag": React / Web Components (experimental) -->