fix(locale): remove countrylist check for translation (#6819)
### Related Ticket(s)

[Services - Masthead/Footer Translations] Translations defaulting to en-US if there is no corresponding country list #6811

### Description

Problem: 
Translations for locales that do not have a corresponding countrylist file default to `us-en`.

Solution:
Remove the countrylist check in the `LocaleAPI` and allow the locale to be passed to the `TranslationAPI`. If a countrylist file for the given locale doesn't exist, make another call to retrieve the default `us-en` countrylist. If in the case there is no translation file for the locale, it will also make another call to the `us-en` translation file.

**Expected Result:**
If adopter tries to get translation for non-existent locale (ie. `eh-ar` as show in the screenshot) it will make a call to both the countrylist and translation file with the locale. Once those fail, it will return the default `us-en` translation and countrylist instead.
<img width="1514" alt="Screen Shot 2021-08-05 at 10 38 03 AM" src="https://user-images.githubusercontent.com/54281166/128370364-90e1ad9b-9620-45bd-9182-c32cc31a99b5.png">


### Changelog

**Changed**

- pass locale from the `LocaleAPI.getLang()` without checking against countrylist
- ensure the `fetchTranslation()` call is made again if the translation call fails

<!-- React and Web Component deploy previews are enabled by default. -->
<!-- To enable additional available deploy previews, apply the following -->
<!-- labels for the corresponding package: -->
<!-- *** "package: services": Services -->
<!-- *** "package: utilities": Utilities -->
<!-- *** "package: styles": Carbon Expressive -->
<!-- *** "RTL": React / Web Components (RTL) -->
<!-- *** "feature flag": React / Web Components (experimental) -->