fix(dotcom-shell): ensure accurate ToC positioning, resolves #6172 (#6880)
### Related Ticket(s)
GH Issue: #6172 
HC JIRA Ticket: https://jsw.ibm.com/browse/HC-2129

### Description

The logic that was in place to position `<dds-table-of-contents>` and `<dds-masthead>` within the `<dds-dotcom-shell>` component on scroll did not account for overscroll in mobile environments. This resulted in miscalculations that progressively pushed the table of contents further down the page if users scrolled up beyond the top of the viewport. 
 
This PR prevents this issue by constraining how far from the top of the viewport we will ever consider the table of contents to be.

https://user-images.githubusercontent.com/12532727/129275283-de91aa35-b4e5-4159-9fe4-787642f80a8f.mov

### Changelog

**Changed**

- Updated `<dds-dotcom-shell>` scroll logic to correctly position `<dds-table-of-contents>` and `<dds-masthead>` components in mobile environments.

<!-- React and Web Component deploy previews are enabled by default. -->
<!-- To enable additional available deploy previews, apply the following -->
<!-- labels for the corresponding package: -->
<!-- *** "package: services": Services -->
<!-- *** "package: utilities": Utilities -->
<!-- *** "package: styles": Carbon Expressive -->
<!-- *** "RTL": React / Web Components (RTL) -->
<!-- *** "feature flag": React / Web Components (experimental) -->