fix(cloud-masthead): prevent masthead shifting on top of page (#6947)
### Related Ticket(s)

[Cloud Masthead] Slight page shift when the Mega Menu is opened on machines where scrollbar is set to visible #6831

### Description

Problem:
The Cloud Masthead shifts when user is on top of the page. The cloud team has applied extra styling to set the position of the masthead to `static` when on top of page. This is so it doesn't overlap their L1. When user scrolls, the `ibm-masthead-sticky ibm-masthead-sticky-showing` v18 classes are applied to the `body` element. This also sets the position of the masthead back to `fixed`. When expanding the megamenu, a `margin-right` needs to be applied to balance out the disappearing scrollbar when position is `fixed`. This `margin-right` is unnecessary when position is `static`.

Solution:
When megamenu is expanded, detect whether the `body` element has the `ibm-masthead-sticky ibm-masthead-sticky-showing` classes applied before setting the `margin-right`.

### Changelog

**Changed**

- only set the `margin-right` for cloud-masthead when the specific v18 classes are applied.

<!-- React and Web Component deploy previews are enabled by default. -->
<!-- To enable additional available deploy previews, apply the following -->
<!-- labels for the corresponding package: -->
<!-- *** "package: services": Services -->
<!-- *** "package: utilities": Utilities -->
<!-- *** "package: styles": Carbon Expressive -->
<!-- *** "RTL": React / Web Components (RTL) -->
<!-- *** "feature flag": React / Web Components (experimental) -->