chore(codeowners): introduce codeowners file (#6988)
### Related Ticket(s)

No related issue

### Description

In order to take advantage of automatic reviewers, a `CODEOWNERS` file needs to be defined. This points the codeowners to the `Carbon for IBM.com members` team. The configuration for the team is to only send the review to the active fireline team, which will be changed in every sprint.

### Changelog

**New**

- `CODEOWNERS` file