fix(search-with-typeahead): Prevent focusOut on button click in safari (#7069)
### Related Ticket(s)
Fixes #5920 

### Description
Safari browsers do not pass focus to buttons when they are clicked. This results in focus leaving the search-with-typeahead component causing it to close. This PR wraps the search button in a `<span tabindex='-1'>` element to catch the focus within the component preventing the focusOut event handler from firing.

### Changelog

**Changed**

- Safari & iOS browsers can now click the magnifying glass icon to search in the masthead