fix(search-with-typeahead): Prevent search form closure on focusout (#7200)
### Related Ticket(s)
fixes #6321
reverts #7069

### Description

This PR removes the logic in the components `_handleFocusOut` method. The method & super call remain to close the dropdown on `focusout`, but closing the form altogether no longer happens. Since the form closure behavior is removed, we can also remove the span tag added in #7069 to prevent safari's a-typical behavior from closing/re-opening the search bar.

### Changelog

**Changed**

- `dds-search-with-typeahead` no longer closes on focusout