fix(leadspace): allow lead space height to grow with content (#7184)
### Related Ticket(s)

#7155

### Description

This PR updates the lead space to grow in height together with the content. Previously the lead space was constructed with fixed heights in mind, but with the spec update it seems some of the `position` rules throughout the component may no longer be required

content should remain visible and the container grows in height. the image is also scaled accordingly:

![image](https://user-images.githubusercontent.com/8265238/133829110-b5936b49-c32f-4a32-9b80-7015e83fed33.png)

Confirm that all lead space variants grow as expected according to the updated component spec

Also confirm that #6323 remains resolved as this PR modifies some changes related to that issue (#6336) 

**question for design:**

in the centered lead space with image, the overlay will need to be adjusted if the content exceeds the min height. how should we handle this?

![image](https://user-images.githubusercontent.com/8265238/133829687-fffaceca-0aee-43c5-9c7d-15c33b8439dd.png)

### Changelog

**New**

- set object-fit and object-position rules depending on lead space size to match current appearance

**Changed**

- use `min-height` rules
- convert leadspace section to flex container

**Removed**

- obsolete positioning rules on leadspace container

<!-- React and Web Component deploy previews are enabled by default. -->
<!-- To enable additional available deploy previews, apply the following -->
<!-- labels for the corresponding package: -->
<!-- *** "package: services": Services -->
<!-- *** "package: utilities": Utilities -->
<!-- *** "package: styles": Carbon Expressive -->
<!-- *** "RTL": React / Web Components (RTL) -->
<!-- *** "feature flag": React / Web Components (experimental) -->