fix(footer): addressing accessibility issues (#7214)
### Related Ticket(s)

[Footer] Resolve a11y issues identified by storybook in web components #7129

### Description

**NOTE**
In order to resolve this a11y issue below, I've set the `hasMainTag` to `true` in the `footer.stories.ts`, as this removes the conflict between the landmarks - the one from the footer itself and the other set by the story container.

<img width="785" alt="Screen Shot 2021-09-21 at 12 58 07" src="https://user-images.githubusercontent.com/42848561/134205261-0fe6c8f9-a4a0-46ae-b67d-5cb138f5712d.png">

The a11y checker is complaining about the `dds-legal-nav-item` as direct children of a ul element. Turns out that the `dds-footer-nav-group` is inside a ul tag as well, and the a11y checker isn't complaining about it. Both the `dds-legal-nav-item` and the `dds-footer-nav-group` has `role="listitem"`, so I'm not really sure what's causing this.
Tried a few approaches to get rid of this serious accessibility issue but none of them worked. 

<img width="546" alt="Screen Shot 2021-09-28 at 16 23 28" src="https://user-images.githubusercontent.com/42848561/135152400-20b14baa-1aa3-48bc-8ba0-52d6eeb23738.png">