fix(MastheadSearch): allow search event handlers to determine method of change (#7365)
### Related Ticket(s)

#7081

### Description

This PR updates the `onSearchNoRedirect` event callback signature which adds the `suggestion` object and `method` from `react-autosuggest` to the returned `event.detail` object.

The internal `onKeyDown` handler is removed so that searches initiated by pressing the <kbd>Enter</kbd> key are not overridden by the `keydown` event handler. If the keydown handler cannot be removed, another option would be to stop the event object from propagating if the `onSearchNoRedirect` handler is called via keyboard input. Since the keydown handler is internal only, it appears to be fine to remove 

There is a temporary storybook example for easier testing

### Changelog

**New**

- `suggestion` and `method` properties on `event.detail` object returned by `onSearchNoRedirect` event

**Changed**

- internal function signatures for `onSearchNoRedirect`

**Removed**

- `MastheadSearch` internal `onKeyDown` handler

<!-- React and Web Component deploy previews are enabled by default. -->
<!-- To enable additional available deploy previews, apply the following -->
<!-- labels for the corresponding package: -->
<!-- *** "package: services": Services -->
<!-- *** "package: utilities": Utilities -->
<!-- *** "package: styles": Carbon Expressive -->
<!-- *** "RTL": React / Web Components (RTL) -->
<!-- *** "feature flag": React / Web Components (experimental) -->