fix(Masthead): resolve React and PropTypes warnings (#7401)
### Related Ticket(s)

#7224

### Description

This PR resolves several React and PropTypes warnings in the Masthead. When viewing the Masthead there should no longer be any warnings in the console

The child elements in `HeaderMenu` always receive a `ref` prop, but it may be improperly passed to the child. Depending on the Masthead variant, the `HeaderMenu` renders the `MegaMenu` or various `HeaderMenuItem`s as descendants. Placing a `ref` on the `MegaMenu` would be invalid since `MegaMenu` is a function component, but `ref`s are valid on `HeaderMenuItem`s

With this PR `MegaMenu` now forwards the ref to its descendant `NavigationGroup` which forwards the ref to a `section` element. Alternatively we can also just not forward a ref and check the `item` type or displayName as a workaround instead https://github.com/carbon-design-system/carbon-for-ibm-dotcom/blob/master/packages/react/src/components/carbon-components-react/UIShell/HeaderMenu.js#L305-L309

### Changelog

**New**

- add unique `key` props to item lists

**Changed**

- rename cloned `ref` prop in `HeaderMenu`
- resolve invalid prop type warnings
- refactor unneeded ternaries and `else-return` statements

<!-- React and Web Component deploy previews are enabled by default. -->
<!-- To enable additional available deploy previews, apply the following -->
<!-- labels for the corresponding package: -->
<!-- *** "package: services": Services -->
<!-- *** "package: utilities": Utilities -->
<!-- *** "RTL": React / Web Components (RTL) -->
<!-- *** "feature flag": React / Web Components (experimental) -->