test(masthead): remove unnecessary percy and cy snapshots (#7495)
### Related Ticket(s)

{{Provide url(s) to the related ticket(s) that this pull request addresses}}

### Description

some e2e tests for the masthead (ie. checking if there are a certain number of menu items or if an element as a url) don't require percy snapshots or cypress screenshots. Having them would add to the number of snapshots and aren't very helpful. This PR removes those methods.

### Changelog

**Removed**

- remove percy and cy snapshots methods from certain tests.

<!-- React and Web Component deploy previews are enabled by default. -->
<!-- To enable additional available deploy previews, apply the following -->
<!-- labels for the corresponding package: -->
<!-- *** "package: services": Services -->
<!-- *** "package: utilities": Utilities -->
<!-- *** "RTL": React / Web Components (RTL) -->
<!-- *** "feature flag": React / Web Components (experimental) -->