feat(masthead): ensure active l1 item is seen on load (#7655)
### Related Ticket(s)
#6473

### Description
This ensures that the selected L1 item will be scrolled to if its out of view upon load. This change also works on the RTL version.

In order to test this out, go to the Masthead L1 story, and under `selected menu item` type `Sed cursus ante dapibus diam`, and open the story in a new tab so the active element updates. You will be able to see the new change upon load. 

### Changelog

**New**

- selected item on the L1 menu will be scrolled to if its out of view

<!-- React and Web Component deploy previews are enabled by default. -->
<!-- To enable additional available deploy previews, apply the following -->
<!-- labels for the corresponding package: -->
<!-- *** "package: services": Services -->
<!-- *** "package: utilities": Utilities -->
<!-- *** "RTL": React / Web Components (RTL) -->
<!-- *** "feature flag": React / Web Components (experimental) -->