fix(content-item-horizontal): use accessible focus order (#7528)
### Related Ticket(s)

Resolves https://github.com/carbon-design-system/carbon-for-ibm-dotcom/issues/6511

### Description

The content-item-horizontal-media component allows adopters to configure which columns the media and text content appear in at desktop viewport widths. This reordering was done with CSS, which created opportunities for the tabbing order to go out of logical sequence. Additionally, the design stipulates that the media content should always appear below the text content at mobile viewport widths.

To fix this, I added conditions to the render method to ensure the markup is printed in the correct logical order. I also added an event listener that triggers a re-render when the screen is resized across the `lg` Carbon breakpoint. This caused some render issues with the footer content which have been (temporarily?) resolved by using [the `shouldUpdate` lifecycle method](https://lit-element.polymer-project.org/guide/lifecycle#shouldupdate).

### Changelog

**Changed**

- Component markup is printed in an order that matches the logical tabbing order.
- Updated CSS to prevent visual regressions.

<!-- React and Web Component deploy previews are enabled by default. -->
<!-- To enable additional available deploy previews, apply the following -->
<!-- labels for the corresponding package: -->
<!-- *** "package: services": Services -->
<!-- *** "package: utilities": Utilities -->
<!-- *** "RTL": React / Web Components (RTL) -->
<!-- *** "feature flag": React / Web Components (experimental) -->