feat(masthead): scroll to active menu item on L0 (#7690)
### Related Ticket(s)
#6473 

### Description
Upon consulting design team, this PR is a follow up to #7655 to add the same scroll to active item functionality to the L0. 

In order to test this out, go to the default Masthead story, and under selected menu item type `Explore more`, and open the story in a new tab so the active element updates. You will be able to see the new change upon load. It may be necessary to shrink the browser width so the L0 is small enough to need the left/right caret pagination functions.

### Changelog

**New**

- added scroll to active element in the L0 menu

<!-- React and Web Component deploy previews are enabled by default. -->
<!-- To enable additional available deploy previews, apply the following -->
<!-- labels for the corresponding package: -->
<!-- *** "package: services": Services -->
<!-- *** "package: utilities": Utilities -->
<!-- *** "RTL": React / Web Components (RTL) -->
<!-- *** "feature flag": React / Web Components (experimental) -->