fix(carousel): prevent scrolling to empty page & ensure scrolling to focused item's page (#7733)
### Related Ticket(s)

Resolves https://github.com/carbon-design-system/carbon-for-ibm-dotcom/issues/7605

### Description

A bug in the focus handler logic forced the carousel to attempt to scroll to page -1 under certain circumstances. This page doesn't exist, of course, and resulted in the carousel scrolling to a blank page.

While I was working on this handler, I found another bug that prevented the first and last carousel items from being scrolled to when they were not in the active carousel page but received focus. Fixing this was a simple matter of 

### Changelog
**Added**
- Added a Video Card Carousel story.

**Changed**
- Updated `DDS Carousel` focus handler logic to correctly identify index of carousel item that has received focus.

**Removed**
- Removed unnecessary conditional that prevented desirable scroll behavior.

<!-- React and Web Component deploy previews are enabled by default. -->
<!-- To enable additional available deploy previews, apply the following -->
<!-- labels for the corresponding package: -->
<!-- *** "package: services": Services -->
<!-- *** "package: utilities": Utilities -->
<!-- *** "RTL": React / Web Components (RTL) -->
<!-- *** "feature flag": React / Web Components (experimental) -->